Initial Commit
In this initial commit, I have set up a new viewer based on the
ExampleMunicipality scene. This viewer also includes a Feature Toggling
system and configuration, including wizard, to show or hide certain
features based on a selection or what is included in the URL's query
parameters.

This will enable organisations without developers to embed the viewer on
their own websites using an IFRAME and by passing along the proper URL
parameters it will instantly show the correct location and feature set