Add new Indicator package with dependencies
The first step into making sure that quality indicators can be exposed for a project area is to create a new package that we can separate into it's own package repo on GitHub.

I already know, thanks to the POC, that I will be needing GeoJSON.net; and luckily, OpenUPM has a package available for us that I could immediately add as dependency.

To make sure this package is well documented and easy to extract; I have also included an initial README with installation instructions.