Load Dossier from GeoShare URL
In this change, I have introduced the object model for a Dossier so that we can actually open it from an external location. Included is a JSONConverter for URIs and (hex) colors; this will make the data model a bit tighter and self-documenting.

I have also made the Dossiers MonoBehaviour -responsible for loading- into a Dossier ScriptableObject that can contain the Dossier data. This will allow for opening multiple dossiers and/or referencing the Dossier from various MonoBehaviours to drive their behaviour.