Introduce Feature button in sidebar and hide by default
I have introduced a new button, and matching icon, for showing the indicator dossier. The button is shown when a dossier loads, and is disabled when the dossier unloads. By disabling it, it will stay in view until it has been opened again.

I have not introduced a Feature SO for this functionality on purpose, as it would show in the SetupWizard and to prevent the feature system from fighting with the opening of the Dossier whether to show the button or not.

This behaviour could change in the future if and once we introduce the opening of Dossiers from within the viewer instead of a shareable link; but for now this is of no consequence.