Show a sidebar with a title and name of the dossier
Here I tied all the work together from the previous commits to make sure that if the toolbutton is enabled for indicators, that a sidebar is shown according to the new designs.

The Sidebar script will be expanded on in future commits, but the general idea is that is should be agnostic from any UI decisions and that it can populate fields with data; such as the name field with the dossier name.