Center camera when clicking on Indicator button
When centering the camera on the dossier area, we should not do that upon loading the scene; but upon opening the dossier using the indicator toolbutton.

This change will split the loading of the Dossier and the project area into two separate activities; where the area is only loaded upon clicking the button.

This should also reduce the number of HTTP calls upon starting the project; so there is a win-win there.