General cleanup and integration for Netherlands3D
There were quite a few files without namespaces, so I added those and moved
the files into a more compact directory structure. There was also an old sample
present in the folders that relied on the Scriptable Object event architecture;
I had to remove that sample and introduce Prefabs and Materials for the
Amsterdam/Netherlands3D repository's sample to keep working.