Dossiers do not open in WebGL
To load the (Geo)JSON files we make use of JSON.net and its converters; but the code stripping of Unity removes all the JsonConverters and the dynamically instantiated classes from GeoJSON.net.

To counter this issue, I have added GeoJSON.net and the whole JSON-Extra's assembly to the `link.xml` instructions, this will prevent unity from stripping these files.

In addition, I have also added the new Indicator values array that can be used in presenting the indicators in the sidebar