Added the ability to show a list of map layers
In this change, I have introduced the UI elements and scripting needed to display a list of maps in the sidebar. The functionality is split into scripts in the package on the one hand, and prefabs in the Twin on the other hand.

The Twin contains a bit of UI that is independent of the package and the package refers to a Toggle Group and Toggles, and is thus agnostic to the structure of the sidebar.

In a future change we could consider adding the UI into the package, but for now this distinction seems to work.