Change address bar when settings change
When the settings are changed, we want to update the address bar so that
people have a deeplink they can copy and share with people.

There was also a bug updating the coordinates when clicking on the
minimap; this worked in the editor, but not in the WebGL build. This has
been corrected, including a bug where the altitude suddenly changed.