Consume Azure Function using API Key
In order to load the dossier and its assets, we need to provide an api key for Azure. These changes will allow you to set an API Key for Indicators and use it to call the API endpoints.

To keep the number of generic dependencies down, I have chosen to add the API key to the indicator configuration; that way you could also read from differing sources.