Dossier doesn't load with latest settings
In the configuration, when you set the dossier id, it starts loading the dossier immediately. But because the base uri and key are set _after_ setting the dossier id, they are not picked up on immediately. By reordering this, and adding a comment to warn devs, it should now work as expected.