Use the DossierListener on the Sidebar ..
The sidebar should show an error message when a dossier fails to open, and a loader while it is opening. This can be dealt with by using a DossierListener and have it enable and disable certain panels when the state of the Dossier System changes.

However, the sidebar needs an initial state depending on the current state of the Dossier system. So I added to the listener that it will invoke the current state's event upon start. This will initialize the sidebar.