Anchor start screen to canvas
To make sure we can always see all elements and have the start screen in view, I have changed it to stretch. Combined with the canvas scaler this should make sure it is always in view.

This does not play nicely with mobile screens or small viewports; but given those are not supported devices now we accept this