Introduce a validation mechanism for functionalities
When a functionality enables, it should check against its validation rules and when an issue arises, it should disable itself again.

We 'let' the functionality enable itself and only after enabling validate it and disable it again. This will allow external forces to interact when enabling, and it will allow for a chain of events that allow for toggles and toolbuttons to disable again instead of silently failing