Link Dome with Terrain functionality
The underground Dome was not linked to any functionality, and it seems this causes a bug in the build because of states of scriptable objects. This caused the need to open a dome twice before it working.

By linking it explicitly to the Terrain functionality, the initialisation routine should work as expected.