Remove dead code - Timeline and Bim Planning import
Teh Bim planning and timeline modules are not on the board to be reintroduced soon, and the original versions may no longer match what we expect from the application.

To this end and help people get acquainted with the application better, we are removing these for now. Should the code be needed again in the future, it can be found in this commit or in the archived Amsterdam/Netherlands3D repository int he Netherlands3D organisation.