First draft to save and load 3d tiles.
This does not work entirely because the flow of data doesn't work as expected. But I want to commit this and push it so that Sam can see what I am working on.

I get a malformed URL error, and I think this is because the URI is not pushed as expected through the flow