Make sure World Origin is placed at NAP 0
The meaning of Origin in the configuration system was originally slightly different from what it was now; and the world origin should be at elevation (NAP) 0 because otherwise the elevation of tiles in Unity would be moved lower and that would effect game objects that assume NAP is at Unity's Y:0.

The Configuration's own Origin should however still contain the elevation because this will be used to position the camera at a certain height.

This fix is however a patch to make configuration work again; the Configuration and Project systems should be reworked so that the configuration is loaded in to the current project and is not a truly separate thing