Fix the Coordinate property -and thus the editor script-
The Origin used to manage the real world Coordinate at which it was located; but this was -partially- moved to the CoordinateSystems class. Because this was partial there was a bug that the Coordinate property of the Origin contained an invalid value.

By making this property a proxy to the CoordinateSystems class' functions; it continues to functions the same as before and now the editor script works again and any system that wants to read this value.