Add type-mapping to data objects using the WCF DataContract attribute.
Newtonsoft JSON.net supports the DataContract attribute, but doesn't support the name and namespace members. By adding our own SerializationBinder, we _can_ pick up on the name and namespace member and use that as a type declaration for in the JSON output.

Through this mechanism, we can ensure that changes in the naming or namespaces of classes do not affect project files stored after this change. As long as the DataContract name and namespace doesn't change the application will always be able to find the associated class.

As a namespace, we have followed LinkedData recommendations by using a URL. Later on we could even add LD schema's at those locations but let's not get ahead of ourselves :)