Ensure correct query parameters escaping, and that we try to use the correct output formats.
In Unity WebGL, when Api Compatibility Level is set to 2.1, there is an issue with EscapeDataString where it doesn't escape spaces. This is fixed in .NET 4.5.1 but enabling the .NET Framework compatibility level in WebGL builds is not recommended due to build sizes and issues with cross-browser compatibility.

Apparently, using Unity's own EscapeURL method on individual query parameter pieces will ensure correct escaping, so that solves the issue.

In addition, the default outputFormat for WFS is "application/json", or "application/json subtype=geojson" for GeoJSON. So I have improved the supported outputFormat determination to take this into account