Add the functionality to import Gltf and Glb files
This commit will add the importer, spawner and layer prefabs to import, persist and load a Gltf/Glb file.

There is however the issue of textures not loading, but I think this has to do with the changes that we did to limit the number of shader variants for Gltfast; and as such stripping the texture loading.