Moved SelectionTools application add-on from PackageStagingArea to BuildingBlocks.
There is already a package called SelectionTools, but there is additional content now in the "PackageStagingArea". I do not expect this to become a package itself, but in line with our architectural principles I have cleaned up the folder and moved it to _BuildingBlocks.

A next step would be to check where and how this content is precisely used - and do another clean up pass or integrate the content into the selection tools package itself