Added the 'Vlooienburg' 3D models
I've added the Vlooienburg 3D models and fixed the materials. Note that the models themselves are pretty broken here and there.. But, we are a viewer, so this is what we got, and this is what they will get back.

The prefab can be found in the folder: Prefabs > Prefabs > ATM_POC > Vlooienburg.prefab.

Just drag / instantiate it into the Main scene and it should snap to it's correct position. (No idea how the floating Origin deals with it though)