It's complicated
Because GetTileSizes() in the TileHandler is called after the zoomable layers are disabled - these layers will no longer provide tile sizes and thus errors occur when you switch zoom level.

By re-invoking GetTileSizes() through reflection -don't hit me!- we compensate for this issue because the newly enabled zoomlayer is now included in the TileSizes list before is it used.