Updated the UI Ticker
UI ticker now only triggers when the mouse is over the selected text. The child objects don't automatically tick any longer. however, when the mouse is over a child with a text that is long enough to 'tick', it also triggers its parent's ticker IF the parent's text is also long enough to 'tick'.