Due to inheritance, an extra invalid request was done
The GeoJsonLayerGameObject class is the base class for our Wfs implementation; but the WFS loads geometric information in the tile handling instead of the gameobject directly.

This caused an extra request -because the GeoJSON flow does it in start- with an invalid parameter, triggering an auth flow in Tygron and causing subssequent requests to be blocked.

By changing the way LayerGameObject initializes, and by doing layer initialization using composition instead of inheritance, we can avoid the extra request and thus the issue