Allow for loading addressable-based assets upon opening a project
In order to re-open addressable based assets, we need to introduce a couple of changes into the flow of the PrefabLibrary. The prefab identifier needs to be known, and searchable, and with it a new layer needs to be instantiated.

The instantiation mechanics have been moved to the LayerGameObjectFactory, previously SpatialObjectFactory, so that the shader fixes are also applied when loaded this way.

With these changes, we do not only provide the ability to load assets -in a backwards-compatible way- from addressables, but also to provide a convenient factory for creating LayerGameObjects -including some post-creation fixes on the name and shaders-.