Introduce the End-To-End (E2E) testing Framework
In Netherlands3D, we want to find a healthy balance between writing automated tests and going too deep into the rabbit hole. In order to do this, we want to have a regression test suite that runs specific tests in the application as if you were a user.

To make sure tests do not become too brittle, the concept of a Page, or in the case of Unity: Scene, Object Model is introduced where we can extract the interfacing with the application away from the actual testing code. Although this can give an initial overhead in creating the first few tests, the POM will ensure that changes in the application won't impact tests individually, but you can update the POM to match the new changes.