Add E2E tests
I have added tests for the following scenario's:

Layers:

- CanOpenLayerPanel
- TerrainLayerShouldBeVisible
- TerrainLayerCanBeMadeInvisible

Minimap:

- MinimapHasAPointer
- MinimapIsShownOnScreen

In the layer tests I had to add a mechanic to execute these tests in sequence because we click through the application here. When don properly, tests should be independent and we should clean up after ourselves; I have chosen a pragmatic approach to keep the tests simple and quick. The best way would probably be to reload the scene before every test, but noone has time for that.

If this will bite me later on, we can make adjustments.