Rework the ColorWheel to be function as intended
The colorwheel only worked correctly if you picked a value using the colorwheel. But when you tried to set an alpha, color value or use hex code, all sorts of funkyness happened.

In this change I have done the following

- Replaced the wheel texture for a Red/Green/Blue color wheel instead of Red/Yellow/Blue
- Rewrote the code in such a way that picking the color will only affect the hue and saturation
- Rewrote the code in such a way that using the opacity slider will only affect opacity
- Rewrote the Darkness, now Color Value, slider to correctly set the Color Value on a scale of 0 to 1 and show the color at Value 1 at max in the picker
- Rewrote the event flow so that whatever you changed, the UI changes with it as expected and you can set components of the color - to connect it to the sliders

The ColorChanged event is now properly emitted at the right time, and a SetColorWithoutNotify was added for initialisation and to keep in mind that sometimes you do not need to dispatch the events