Authentication failed when loading project with 2 datasources
In the KeyVault, there was a coroutineRunner that was killed off the second time you tried to authenticate a service. This meant that when a project has multiple services, it will load them in rapid succession, killing of prior coroutineRunners and thus the coroutines searching for the authentication.

By creating a dedicated coroutineRunner that is long lived, we can start multiple coroutines without fear of them vanishing :)