Rewrote StyleSelectorParser to StyleResolver - which is now capable of resolving the whole style for a feature.
Rewrote the Expressions to use a new ExpressionValue object instead of IConvertible to remove the weird type juggling that I had to do with the Array object (and that is now gone) and it makes the values more clear and typestrict.

Started refactoring the implementation for CartesianTiles to make use of the new objects and structure.