Fixed bug with Min expression
The min expression returned the wrong value - rhs.Value instead of rhsValue - and the tests did not check the outcome.

I have fixed this bug and altered all ExpressionEvaluator tests to include a check whether the result of the evaluation is an ExpressionValue object. This will verify that the outcome is properly resolved to a value