Refactored code to use Uxios' QueryParameters object and QueryString conversion class
In Uxios v0.4.0 there has been a ton of memory allocation improvements when it comes to parsing querystrings. By leveraging these, we reduce the amount of required memory when loading large datasets and improve performance at the same time.

Not included in this commit is the actual upgrade since I tested against the git package directly. Once the build of 0.4.0 is done I will commit the package changes