Interpret Datasets with multiple distributions as a DataSet in our catalog
Datasets are catalog records that represent multiple distributions of the same data in various ways. For example, a trees endpoint with a textured and untextured endpoint, or parking spots with a WFS and a WMS representation.

When we encounter a DataSet with one distribution, we convert that into a singular record as it doesn't make sense to go a level deeper for something with a singular representation - but when there are multiple distributions, an extra folder level is needed to group these together.

This also demonstrates the power of the Record Strategies as each strategy could even create a tree of objects by re-using the features from the InMemoryCatalog