Restrict support to Crawleable and Searchable catalogs
An OGC API Features reference is commonly a reference to another OGC API. This is a bit of a misnomer that can happen. OGC API Records supports using this as a catalog according to the Local Resources Catalog deployment type, but for the testbed we do not support this yet.

As such, in this commit I have added the restriction that a catalog should have itemType: record in their collections endpoint (which is required according to the spec for crawlable and searchable deployment types)