Imported code from https://github.com/Netherlands3D/SimpleJSON
Because https://github.com/Netherlands3D/SimpleJSON is a fork of
https://github.com/Bunny83/SimpleJSON and we have made
way too many changes to it since then - I had to copy the code and
accept the loss of history.

If you want to review the commit history, go to the archived repository
and look through that.

This code is taken as is, meaning that it should be identical to that
repository.