Re-orient decalprojectors after shift
The origin shifting code re-orients the rotation to match the curvature of the CRS. This does mean that the transform.rotation is altered and since decalprojectors have a custom override of 90 on the x-axis, this override is undone when shifting.

This causes tiles to be invisible because the projector projects it on the wrong axis.

This commit is a quickfix that can be fixed in 1904 so that it works now