Prevent crashes with WMS endpoints that error
Sometimes, endpoints error because authentication hasn't happened yet, has expired or it is flaky.

When this happens, the application crashes because it tries to access memory that has been cleared.

The cause is the capturing of a Tile reference lambda when downloading it as the Tile object is released/destroyed but the download is async and finishes afterwards. This makes the pointer to the tile invalid and accessing is causes an error in the build due to the way WASM/Unity memory accessors work.

In addition I have

- moved projector specific code to the projector, eliminating 2 GetComponent calls in the tile loading path
- Removed the "compress" action and ensured that the texture is loaded non-readable

The latter is an important memory improvement since compress may not even yield any advantages but the texture will take up twice the space during loading because it needs to be readable. By removing the compress step we prevent memory issues because of concurrent loading and simplify the code