Disable tooltips for the credentials and fix help url
All buttons in the credential menu have a tooltip saying: Add to terrein. This is incorrect, and one can even debate whether a tooltip is needed there - as such I have disabled tooltips for now in the credential menu except for the help icon.

In addition: I have fixed the url in the help icon to point to the right page in the documentation